Add team names to cfg schema
This fixes StrictYAML throwing while validating the config data.

Because we're using a StrictYAML schema to validate all config var types
(because we use cfg vars for things such as lists/numeric types/etc, and
we wanna ensure they aren't malformed when evaluated from "text form"),
the new team name config var strings also need to be explicitly specified
in that schema in order to pass the StrictYAML loading check.